requirements.txt
asgiref==3.4.1
astroid==2.7.2
dj-database-url==0.5.0
Django==3.2.6
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
drf-writable-nested==0.6.3
factory-boy==3.2.0
Faker==8.12.1
gunicorn==20.0.4
isort==5.9.3
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.3.2
platformdirs==2.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
whitenoise==5.3.0
wrapt==1.12.1
